Exercise for delete button and authority for login

+++ vidly-app/src/components/moviesTable.jsx
@@ -2,6 +2,7 @@ import React, { Component } from 'react';
 import { Link } from 'react-router-dom';
 import Table from './common/table';
 import Like from './common/like';
+import auth from '../services/authService';
 
 class MoviesTable extends Component {
   columns = [
@@ -21,14 +22,21 @@ class MoviesTable extends Component {
     },
     {
       key: 'delete',
-      content: movie => (
-        <button
-          onClick={() => this.props.onDelete(movie)}
-          className='btn btn-danger btn-sm'
-        >
-          Delete
-        </button>
-      ),
+      content: movie => {
+        const user = auth.getCurrentUser();
+        if (user && user.isAdmin) {
+          return (
+            <button
+              onClick={() => this.props.onDelete(movie)}
+              className='btn btn-danger btn-sm'
+            >
+              Delete
+            </button>
+          );
+        } else {
+          return null;
+        }
+      },
     },
   ];
